Setup django project
Models and some serializers

It functions

Need to write the rest of the serializers and endpoints, figure out how to talk
to postgres, and confirm that validation works the way I want it to.

The rest of the serializers and endpoints™

This should be the basics. No guarantee of correctness. Need to check on
validation and what have you, yada yada.

Generics are cool

Basic admin or naw authentication

Fine tune permissions a lil bit?

IDK if this is correct, but I think I'm starting to understand how auth
works.

I need to go to bed.

Ops scaffolding

More ops scaffolding

A lot of this is going to have to change because I am going to bring in
nginx

Ops scaffolding 3: The epic conclusion

I added an Nginx server so that we could serve static files and gave a
best guess at configuring CSRF trusted origins.

Update README.md

Fun badges

Remove boilerplate

Andy Comments

black?

Make sure debug mode is OFF!!!

Whoops missed a 'lint'

Remove more boilerplate

Andrew Comments Part 1